+++ tests/client_headers.rs
@@ -80,7 +80,7 @@ fn test_headers() {
 
     // This read should succeed even though we already initiated a close
     let message = client_handler.read().unwrap();
-    assert_eq!(message.into_data(), b"Hello WebSocket");
+    assert_eq!(message.into_data(), b"Hello WebSocket"[..]);
 
     assert!(client_handler.read().unwrap().is_close()); // receive acknowledgement