+++ tests/receive_after_init_close.rs
@@ -46,7 +46,7 @@ fn test_receive_after_init_close() {
 
     // This read should succeed even though we already initiated a close
     let message = client_handler.read().unwrap();
-    assert_eq!(message.into_data(), b"Hello WebSocket");
+    assert_eq!(message.into_data(), b"Hello WebSocket"[..]);
 
     assert!(client_handler.read().unwrap().is_close()); // receive acknowledgement